Early feedback updates and refactoring
