+++ packages/@okta/i18n/src/properties/login.properties
@@ -1657,7 +1657,6 @@ idx.error.code.access_denied.device_assurance.remediation.chrome.chromeos.key_tr
 # End user remediation messages for Android Zero Trust signals
 idx.error.code.access_denied.device_assurance.remediation.android.zero.trust.play_protect_required = Please ensure Google Play Protect is enabled and that a scan has been performed, then try accessing this application again.
 idx.error.code.access_denied.device_assurance.remediation.android.zero.trust.play_protect_scan_threshold = There is a problem with your Google Play Protect system scan. Please open Google Play and take action on the warnings displayed.
-idx.error.code.access_denied.device_assurance.remediation.android.zero.trust.device_integrity_not_met = Device integrity not met.
 idx.error.code.access_denied.device_assurance.remediation.android.zero.trust.version_update_available = Your device has an OS update available. Please update your device then log in again.
 idx.error.code.access_denied.device_assurance.remediation.android.zero.trust.disable_usb_debugging = Turn off USB debugging from Developer Options
 idx.error.code.access_denied.device_assurance.remediation.android.zero.trust.lock_screen_complexity = Your lock screen doesn't meet complexity requirements.
