+++ src/v3/src/util/formUtils.ts
@@ -10,7 +10,6 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-import fetch from 'cross-fetch';
 import {
   IdxAuthenticator,
   IdxMessage, IdxRemediation, IdxTransaction, NextStep,
@@ -18,7 +17,6 @@ import {
 
 import { LanguageCode } from '../../../types';
 import IDP from '../../../util/IDP';
-import Logger from '../../../util/Logger';
 import TimeUtil from '../../../util/TimeUtil';
 import Util from '../../../util/Util';
 import {
@@ -39,8 +37,7 @@ import {
 } from '../types';
 import { idpIconMap } from './idpIconMap';
 import { loc } from './locUtil';
-
-const ADP_INSTALL_FALLBACK_REMEDIATION_KEY = 'idx.error.code.access_denied.device_assurance.remediation.android.zero.trust.android_device_policy_app_required_manual_install';
+import { probeLoopbackAndExecute } from './probeLoopbackAndExecute';
 
 export type PhoneVerificationStep = typeof IDX_STEP.CHALLENGE_AUTHENTICATOR
 | typeof IDX_STEP.AUTHENTICATOR_VERIFICATION_DATA;
@@ -313,11 +310,10 @@ const buildEnduserRemediationWidgetMessageLink = (
       url: links[0].url,
       label: message,
     };
-  } else if (deviceRemediation.action) {
+  } else if (deviceRemediation.remediationType === 'LOOPBACK' && deviceRemediation.action) {
     return {
       label: message,
       onClick: () => {
-        console.log('clicked the device remediation link button, triggering loopback...');
         probeLoopbackAndExecute(deviceRemediation);
       },
       dataSe: deviceRemediation.action,
@@ -524,145 +520,4 @@ export const buildPhoneVerificationSubtitleElement = (
   };
 
   return subtitleElement;
-};
-
-type RequestOptions = {
-  url: string;
-  timeout: number;
-  method: 'GET' | 'POST';
-  data?: string;
-};
-
-/**
-* Temporary request client can be removed if this functionality is added
-* to auth-js library.
-* @see https://oktainc.atlassian.net/browse/OKTA-561852
-* @returns Promise<Response>
-*/
-export const makeRequest = async ({
- url, timeout, method, data,
-}: RequestOptions) => {
- // Modern browsers support AbortController, so use it
- if (window?.AbortController) {
-   const controller = new AbortController();
-   const id = setTimeout(() => controller.abort(), timeout);
-
-   const response = await fetch(url, {
-     method,
-     headers: {
-       'Content-Type': 'application/json',
-     },
-     ...(typeof data === 'string' ? { body: data } : {}),
-     signal: controller.signal,
-   });
-
-   clearTimeout(id);
-
-   return response;
- }
-
- // IE11 does not support AbortController, so use an alternate
- // timeout mechanism
- const responsePromise = fetch(url, {
-   method,
-   headers: {
-     'Content-Type': 'application/json',
-   },
-   ...(typeof data === 'string' ? { body: data } : {}),
- });
- const timeoutPromise = new Promise<Response>((_, reject) => {
-   setTimeout(() => {
-     const abortError = new Error('Aborted');
-     abortError.name = 'AbortError';
-     reject(abortError);
-   }, timeout);
- });
-
- return Promise.race([responsePromise, timeoutPromise]);
-};
-
-const getMessage = (fallback: DeviceRemediation['fallback']) => {
-  switch (fallback.message?.i18n?.key) {
-    case ADP_INSTALL_FALLBACK_REMEDIATION_KEY:
-      // ADP fallback message has additional formatting
-      return loc(
-        fallback.message.i18n.key,
-        'login',
-        fallback.message.i18n.params,
-        { $1: { element: 'span', attributes: { class: 'strong nowrap' } } },
-      );
-    default:
-      return loc(fallback.message!.i18n.key, 'login', fallback.message?.i18n.params);
-  }
-};
-
-const executeDeviceRemediationFallback = (fallback: DeviceRemediation['fallback'], action: string) => {
-  switch (fallback.type) {
-    case 'APP_LINK':
-      // If/when loopback fails auto launch app link to install
-      Util.executeOnVisiblePage(() => {
-        Util.redirectWithFormGet(fallback.href);
-      });
-      break;
-    case 'MESSAGE':
-      // display updated message in place
-      const elements = document.querySelectorAll(`[data-se="${action}"]`);
-      if (elements?.[0]) {
-        elements[0].outerHTML = getMessage(fallback);
-      }
-      break;
-    default: // Do nothing
-  }
-};
-
-export const probeLoopbackAndExecute = async (deviceRemediation: DeviceRemediation) => {
-  let remediationSucceeded = false;
-
-  const baseUrls = deviceRemediation.ports.map((port) => `${deviceRemediation.domain}:${port}`);
-
-  for (const baseUrl of baseUrls) {
-    try {
-      // probe the port
-      const probeResponse = await makeRequest({
-        method: 'GET',
-        timeout: deviceRemediation.probeTimeoutMillis,
-        url: `${baseUrl}/probe`,
-      });
-
-      if (!probeResponse.ok) {
-        Logger.error(`Loopback is not listening on: ${baseUrl}.`);
-        continue;
-      }
-
-      // attempt the remediation request with found port
-      const remediationResponse = await makeRequest({
-        url: `${baseUrl}/${deviceRemediation.action}`,
-        method: 'GET',
-        timeout: 300_000,
-      });
-
-      if (!remediationResponse.ok) {
-        if (remediationResponse.status !== 503) {
-          // return immediately
-          return;
-        }
-
-        // try the next domain:port
-        continue;
-      }
-      // remediation response was a 2xx, end probing
-      remediationSucceeded = true;
-      break;
-    } catch (e) {
-      // only for unexpected error conditions (e.g. fetch throws an error)
-      Logger.error(`Something unexpected happened while we were checking url ${baseUrl}`);
-    }
-  }
-
-  if (!remediationSucceeded) {
-    // if remediation request was not made after probe, utilize fallback mechanism
-    Logger.error('No available ports. Loopback server failed, triggering fallback.');
-
-    executeDeviceRemediationFallback(deviceRemediation.fallback, deviceRemediation.action);
-  }
 };
