+++ src/v3/src/util/index.ts
@@ -41,7 +41,9 @@ export * from './isPasswordRecovery';
 export * from './isPollingStep';
 export * from './languageUtils';
 export * from './locUtil';
+export * from './makeRequest';
 export * from './passwordUtils';
+export * from './probeLoopbackAndExecute';
 export * from './punctuate';
 export * from './removeFieldLevelMessages';
 export * from './rendererUtils';
